requirements.txt
pandas>=1.1.0
snowflake-connector-python>=2.3.0
openpyxl>=3.0.0
Pillow>=8.0.0
ttkthemes>=3.0.0
python-dateutil>=2.8.0
pytz>=2020.0
PyYAML>=6.0.0